requirements.txt
pyobjc
python-dotenv